Added the Conversion of Counter_Price and the Price_Code and the calculation based on the Markup

+++ database/migrations/2024_03_02_041625_create_products_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
                 $table->string('model')->nullable();
                 $table->string('price_code')->nullable();
                 $table->decimal('markup', 8, 2)->nullable(); //Percentage
-                $table->integer('counter_price')->nullable(); //The Output for the Price_Code * (percentage + 1)
+                $table->decimal('counter_price', 8, 2)->nullable(); //The Output for the Price_Code * (percentage + 1)
                 $table->integer('stock')->nullable();
                 $table->timestamps();
                 $table->softDeletes();
